backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.8
motor==3.3.2
pymongo==4.5.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
mtcnn==0.1.1
scikit-learn==1.3.0
Pillow==10.0.0